requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
python-dotenv==1.0.1
httpx==0.27.0
gunicorn
anthropic==0.25.2
google-generativeai==0.7.0
openai==1.35.13